Improved the pretty print functionality.
Now it is not required the print.yml, it will set background_color=white and foreground_color=black.
But it can be used optionally to override other properties.

Fixed: Show only print in dashboard view.
Fixed: Not able to print time intervals.
Fixed: Not able to print detailed view.

Conflicts:
	views/layout.erb