Add a script to validate dashboards.
This script validate that all the categories contains a description (the `dash.yaml` file), and that all the `.graph` files can be parsed with `GraphiteGraph`.

When a new graph is created, if there's a syntax error in the graph description, the application will be unable to display it.  If the dashboards are stored in some kind of repository, before pulling them to the server, or before pushing them, this script can be run to validate that the graphs can be loaded by `GraphiteGraph` and that all categories contains the dash description (e.g: a git hook, Jenkin's job, etc.)